Refactoring of filter and sort functions to data utils file
